add Obj3BP constructor from linear vertices array, upgrades on PG and BP

+++ include/skyboxPG.hpp
@@ -17,7 +17,7 @@ public:
 	SkyboxPG(std::string vs_file, std::string fs_file);
 	~SkyboxPG();
 
-	virtual void	renderObjects(list<Object*>& list, Cam& cam, uint8_t flags = 0);
+	virtual void	renderObjects(list<Object*>& list, Cam& cam, unsigned int flags = 0);
 	virtual void	render(Object& object, Math::Matrix4 VPmatrix) const;
 	/*
 		https://cpp.developpez.com/faq/cpp/?page=Les-fonctions-membres-virtuelles
